scripts/requirements.txt
codeowners==0.6.0
ruamel.yaml==0.18.5
slack_sdk
